requirements.txt
# Mail Merge SaaS Requirements - Pandoc PDF Generation
# Install with: pip install -r requirements.txt

# Core web framework
flask==2.3.3
werkzeug==2.3.7

# Document processing (pure Python)
python-docx==0.8.11
openpyxl==3.1.2

# Template engine
jinja2==3.1.2

# Production server
gunicorn==21.2.0